moved create user action to own directory * needed for gazelle to generate separate build rule for each action
