+++ cmd/serve/BUILD.bazel
@@ -9,14 +9,14 @@ load("@rules_pkg//pkg:tar.bzl", "pkg_tar")
 go_library(
     name = "serve_lib",
     srcs = ["main.go"],
-    deps = [
-        "//handler",
-        "@com_github_jmoiron_sqlx//:sqlx",
-        "@com_github_lib_pq//:pq",
-        ],
     embedsrcs = [":web"],
     importpath = "github.com/sunb26/xat/cmd/serve",
     visibility = ["//visibility:private"],
+    deps = [
+        "//handler/create_user",
+        "@com_github_jmoiron_sqlx//:sqlx",
+        "@com_github_lib_pq//:pq",
+    ],
 )
 
 copy_to_directory(
