+++ handler/create_user/BUILD.bazel
@@ -1,11 +1,9 @@
 load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "handler",
-    srcs = [
-        "create_user.go",
-    ],
-    importpath = "github.com/sunb26/xat/handler",
+    name = "create_user",
+    srcs = ["create_user.go"],
+    importpath = "github.com/sunb26/xat/handler/create_user",
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_jmoiron_sqlx//:sqlx",
